Temporarily disable masking
I need to fix a bug where a student re-visits
a problem that they had solved without masking enabled.
That case does not currently work correclty.
It's simplest to turn off masking now, and re-enable
it with the whole fix.

This change is one line to turn off masking, and a bunch
of lines to make the tests consistent.